requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
idna==2.6
pendulum==1.4.1
pyTelegramBotAPI==3.5.2
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.3
pytzdata==2018.3
requests==2.18.4
six==1.11.0
tzlocal==1.5.1
urllib3==1.22
